Enable gristle_differ to use the python sort
gristle_differ was only using the unix-based sort library in
datagristle, this change enables it to use the python-based sort as
well.

The decision on which sort method to use is based on the dialect - any
non-quoted files will use the unix sort, and any form of quoting will
use the python sort.

The rational is that the python sort is safer, but slower.  The unix
sort should be fine for non-quoted files.