gristle_profiler: add column_name referencing
This allows user to restrict analysis to a column specified either by
offset (original functionality) or now by name.  The name comes from the
header.

BREAKING: this renames the option column_number to simply column - since
it contains names as well as offsets.  If a user provides the option
column_number it will be caught, the pgm will abort, and a message will
describe the new option name to use.