common.abort - write to stderr and less quoting
Write to stderr instead of stdout for err messages.  This shouldn't be
much of a breaking change for anyone.

Also, stop using repr() on text string unless necessary: it causes text
to get surrounded with quotes.  This change will ensure that pure-string
fields will be printed normally, and objects will still get a repr().