validator: remove validictory & cletus modules
This change completes the removal of validictory from datagristle.

And it removes the cletus suppressions module from gristle_validator.
This module isn't that useful - it simply allows one to temporarily
suppress the behavior of a frequently scheduled job while still leaving
it in the schedule.  Only slightly useful, so getting rid of it.