Fix config file & env boolean logic
The problem is that file configs and envvars don't handle bools the same
way that the cli does: rather than always apply the const value, the
user can set them to None, False, '', etc.

This change ensures that the only value accepted from envvars or config
files for booleans are true, True, t, or 1.  Anything else will result
in the program aborting. And then the const value is applied, not the
given value.