refactor(csvhelper): consolidate sloppy functions
We previously had a number of functions related to determining the csv
dialect to use.  We had these because we want to autodetect (sniff) this
from files, but also let the user provide overrides, and finally we
needed some defaulting of final missing bits and validation.  The result
was a PITA to use and difficult to understand.

The new process replaces all that with a simple Dialect class and
BuildDialect class.

This commit is huge because I wanted to see what it was going to look
like.  Need to follow this up with a little clean-up - around type,
comments, tests.