fix (gristle_differ) noisy sqlalchmy logging
Prevent sqlalchemy from spewing noisy logs to the user.

This redirects them to /tmp, with an iso8601 format - at least for
gristle_differ.  It also still lets any CRITICAL logs through to
the user.

It would be a good idea to expand upon this later as a
general method of writing logs.